refactor: rename folders and files, update import/export paths
